refactor(optimizers): optimizer/optimizers
