+++ src/query/service/tests/it/sql/planner/optimizer/agg_index_query_rewrite.rs
@@ -28,8 +28,8 @@ use databend_common_expression::TableField;
 use databend_common_expression::TableSchemaRefExt;
 use databend_common_meta_app::schema::CreateOption;
 use databend_common_sql::optimizer::ir::SExpr;
+use databend_common_sql::optimizer::optimizers::RecursiveOptimizer;
 use databend_common_sql::optimizer::OptimizerContext;
-use databend_common_sql::optimizer::RecursiveOptimizer;
 use databend_common_sql::optimizer::RuleID;
 use databend_common_sql::optimizer::DEFAULT_REWRITE_RULES;
 use databend_common_sql::plans::AggIndexInfo;
