+++ src/query/sql/src/planner/optimizer/optimizer.rs
@@ -25,11 +25,9 @@ use log::info;
 use crate::binder::target_probe;
 use crate::binder::MutationStrategy;
 use crate::binder::MutationType;
-use crate::optimizer::cascades::CascadesOptimizer;
 use crate::optimizer::distributed::optimize_distributed_query;
 use crate::optimizer::distributed::BroadcastToShuffleOptimizer;
 use crate::optimizer::distributed::SortAndLimitPushDownOptimizer;
-use crate::optimizer::hyper_dp::DPhpy;
 use crate::optimizer::ir::Memo;
 use crate::optimizer::ir::SExpr;
 use crate::optimizer::operator::DeduplicateJoinConditionOptimizer;
@@ -38,11 +36,12 @@ use crate::optimizer::operator::RuleNormalizeAggregateOptimizer;
 use crate::optimizer::operator::RuleStatsAggregateOptimizer;
 use crate::optimizer::operator::SingleToInnerOptimizer;
 use crate::optimizer::operator::SubqueryRewriter;
-use crate::optimizer::rule::TransformResult;
+use crate::optimizer::optimizers::CascadesOptimizer;
+use crate::optimizer::optimizers::DPhpy;
+use crate::optimizer::optimizers::RecursiveOptimizer;
 use crate::optimizer::statistics::CollectStatisticsOptimizer;
 use crate::optimizer::util::contains_local_table_scan;
 use crate::optimizer::util::contains_warehouse_table_scan;
-use crate::optimizer::RuleFactory;
 use crate::optimizer::RuleID;
 use crate::optimizer::DEFAULT_REWRITE_RULES;
 use crate::planner::QueryExecutor;
@@ -120,72 +119,6 @@ impl OptimizerContext {
     pub fn with_max_push_down_limit(mut self, max_push_down_limit: usize) -> Self {
         self.max_push_down_limit = max_push_down_limit;
         self
-    }
-}
-
-/// A recursive optimizer that will apply the given rules recursively.
-/// It will keep applying the rules on the substituted expression
-/// until no more rules can be applied.
-pub struct RecursiveOptimizer<'a> {
-    ctx: &'a OptimizerContext,
-    rules: &'static [RuleID],
-}
-
-impl<'a> RecursiveOptimizer<'a> {
-    pub fn new(rules: &'static [RuleID], ctx: &'a OptimizerContext) -> Self {
-        Self { ctx, rules }
-    }
-
-    /// Run the optimizer on the given expression.
-    #[recursive::recursive]
-    pub fn run(&self, s_expr: &SExpr) -> Result<SExpr> {
-        self.optimize_expression(s_expr)
-    }
-
-    #[recursive::recursive]
-    fn optimize_expression(&self, s_expr: &SExpr) -> Result<SExpr> {
-        let mut optimized_children = Vec::with_capacity(s_expr.arity());
-        let mut children_changed = false;
-        for expr in s_expr.children() {
-            let optimized_child = self.run(expr)?;
-            if !optimized_child.eq(expr) {
-                children_changed = true;
-            }
-            optimized_children.push(Arc::new(optimized_child));
-        }
-        let mut optimized_expr = s_expr.clone();
-        if children_changed {
-            optimized_expr = s_expr.replace_children(optimized_children);
-        }
-
-        let result = self.apply_transform_rules(&optimized_expr, self.rules)?;
-
-        Ok(result)
-    }
-
-    fn apply_transform_rules(&self, s_expr: &SExpr, rules: &[RuleID]) -> Result<SExpr> {
-        let mut s_expr = s_expr.clone();
-        for rule_id in rules {
-            let rule = RuleFactory::create_rule(*rule_id, self.ctx.clone())?;
-            let mut state = TransformResult::new();
-            if rule
-                .matchers()
-                .iter()
-                .any(|matcher| matcher.matches(&s_expr))
-                && !s_expr.applied_rule(&rule.id())
-            {
-                s_expr.set_applied_rule(&rule.id());
-                rule.apply(&s_expr, &mut state)?;
-                if !state.results().is_empty() {
-                    // Recursive optimize the result
-                    let result = &state.results()[0];
-                    let optimized_result = self.optimize_expression(result)?;
-                    return Ok(optimized_result);
-                }
-            }
-        }
-
-        Ok(s_expr.clone())
     }
 }
