+++ src/query/sql/src/planner/optimizer/optimizers/cascades/cascade.rs
@@ -19,16 +19,16 @@ use databend_common_exception::Result;
 use log::debug;
 
 use super::explore_rules::get_explore_rule_set;
-use crate::optimizer::cascades::scheduler::Scheduler;
-use crate::optimizer::cascades::scheduler::DEFAULT_TASK_LIMIT;
-use crate::optimizer::cascades::tasks::OptimizeGroupTask;
-use crate::optimizer::cascades::tasks::Task;
 use crate::optimizer::cost::CostModel;
 use crate::optimizer::cost::DefaultCostModel;
 use crate::optimizer::ir::Distribution;
 use crate::optimizer::ir::Memo;
 use crate::optimizer::ir::RequiredProperty;
 use crate::optimizer::ir::SExpr;
+use crate::optimizer::optimizers::cascades::OptimizeGroupTask;
+use crate::optimizer::optimizers::cascades::Scheduler;
+use crate::optimizer::optimizers::cascades::Task;
+use crate::optimizer::optimizers::cascades::DEFAULT_TASK_LIMIT;
 use crate::optimizer::rule::TransformResult;
 use crate::optimizer::OptimizerContext;
 use crate::optimizer::RuleSet;
