+++ src/query/sql/src/planner/optimizer/optimizers/hyper_dp/dphyp.rs
@@ -21,13 +21,13 @@ use databend_common_catalog::table_context::TableContext;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 
-use crate::optimizer::hyper_dp::join_node::JoinNode;
-use crate::optimizer::hyper_dp::join_relation::JoinRelation;
-use crate::optimizer::hyper_dp::join_relation::RelationSetTree;
-use crate::optimizer::hyper_dp::query_graph::QueryGraph;
-use crate::optimizer::hyper_dp::util::intersect;
-use crate::optimizer::hyper_dp::util::union;
 use crate::optimizer::ir::SExpr;
+use crate::optimizer::optimizers::hyper_dp::util::intersect;
+use crate::optimizer::optimizers::hyper_dp::util::union;
+use crate::optimizer::optimizers::hyper_dp::JoinNode;
+use crate::optimizer::optimizers::hyper_dp::JoinRelation;
+use crate::optimizer::optimizers::hyper_dp::QueryGraph;
+use crate::optimizer::optimizers::hyper_dp::RelationSetTree;
 use crate::optimizer::rule::TransformResult;
 use crate::optimizer::OptimizerContext;
 use crate::optimizer::RuleFactory;
