+++ src/query/sql/src/planner/optimizer/optimizers/hyper_dp/join_node.rs
@@ -16,9 +16,9 @@ use std::sync::Arc;
 
 use databend_common_exception::Result;
 
-use crate::optimizer::hyper_dp::join_relation::JoinRelation;
 use crate::optimizer::ir::RelExpr;
 use crate::optimizer::ir::SExpr;
+use crate::optimizer::optimizers::hyper_dp::JoinRelation;
 use crate::plans::Join;
 use crate::plans::JoinEquiCondition;
 use crate::plans::JoinType;
