+++ src/query/sql/src/planner/optimizer/optimizers/hyper_dp/query_graph.rs
@@ -17,7 +17,7 @@ use std::collections::HashSet;
 
 use databend_common_exception::Result;
 
-use crate::optimizer::hyper_dp::util::is_subset;
+use crate::optimizer::optimizers::hyper_dp::util::is_subset;
 use crate::IndexType;
 use crate::ScalarExpr;
