+++ src/query/sql/src/planner/optimizer/optimizers/hyper_dp/util.rs
