+++ src/query/sql/src/planner/optimizer/optimizers/recursive/mod.rs
@@ -12,25 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod cost;
-mod distributed;
-pub mod ir;
-pub mod operator;
-#[allow(clippy::module_inception)]
 mod optimizer;
-pub mod optimizers;
-pub mod rule;
-mod statistics;
-mod util;
 
-mod dynamic_sample;
-
-pub use optimizer::optimize;
-pub use optimizer::optimize_query;
-pub use optimizer::OptimizerContext;
-pub use rule::agg_index;
-pub use rule::RuleFactory;
-pub use rule::RuleID;
-pub use rule::RuleSet;
-pub use rule::DEFAULT_REWRITE_RULES;
-pub use util::contains_local_table_scan;
+pub use optimizer::RecursiveOptimizer;
